Rename Tests.Domain.Primitives to Tests.Domain.Shared for Domain consistency
Rename Tests.Domain.Primitives to Tests.Domain.Shared for Domain consistency
