+++ account-management/tests/Tests/Domain/Shared/EntityTests.cs
@@ -3,7 +3,7 @@ using JetBrains.Annotations;
 using PlatformPlatform.AccountManagement.Domain.Shared;
 using Xunit;
 
-namespace PlatformPlatform.AccountManagement.Tests.Domain.Primitives;
+namespace PlatformPlatform.AccountManagement.Tests.Domain.Shared;
 
 public static class EntityTests
 {
